list of changes:  - Some tests are moved from MeshTests.h to ApplicationTests.h  - Duplicating fields are deleted from Application class
